refactor(testing): decompose test_round_trip.py into round_trip_* siblings
MAJ-04 module 6/6: gnn/testing/test_round_trip.py 2214 -> 1356 lines
(facade).

- round_trip_config.py: six test-configuration dictionaries (leaf)
- round_trip_results.py: RoundTripResult/ComprehensiveTestReport dataclasses
- round_trip_availability.py: serializer/validator try-import probe
  (single owner of GNN_AVAILABLE/CROSS_FORMAT_AVAILABLE and the 30+
  parser/serializer/validator names; re-imports LOGGING_CONFIG)
- round_trip_markdown_parser.py: _DirectMarkdownParser (leaf)
- round_trip_comparison.py: RoundTripComparisonMixin (_compare_* +
  _compute_model_checksum + _test_cross_format_consistency)
- round_trip_report.py: RoundTripReportMixin (generate_report)
- test_round_trip.py: facade - GNNRoundTripTester(mixins) keeps the
  availability side effects (sys.setrecursionlimit, sys.path.insert,
  logging.basicConfig) verbatim, the tester core, _get_file_extension,
  TestGNNRoundTrip, and the __main__ driver; re-exports all 47 moved
  names incl. every parser/serializer/validator binding

Pattern: execute/processor.py mechanical split; class methods move via
mixins to preserve byte-identity. Siblings avoid test_* names so
explicit-path pytest collection semantics are unchanged. acyclic DAG:
config/results/availability <- {markdown_parser, comparison, report} <- facade.

2172 lines verified byte-identical (moved + kept); only import headers
(ruff-hoisted), the class base list, and TYPE_CHECKING stubs differ.

Verified: ruff 0 (repo-wide); mypy 0 (621 files); facade contract 14
names + the three lazy consumers (schema_validator/validator,
round_trip_strategy, test_integration) importable; targeted pytest
src/gnn/testing/test_round_trip.py: 3 passed, 1 pre-existing data
failure (test_reference_file_validation needs the long-deleted
input/gnn_files/actinf_pomdp_agent.md primary reference; test is
outside the command of record per pytest.ini testpaths=tests/).
test_comprehensive_round_trip passes end-to-end through every moved
component.
